+++ popi-reservation-service/src/main/java/com/lgcns/PopiReservationServiceApplication.java
@@ -1,4 +1,4 @@
-package com.lgcns.popireservationservice;
+package com.lgcns;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
