+++ popi-reservation-service/src/test/java/com/lgcns/PopiReservationServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.lgcns.popireservationservice;
+package com.lgcns;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
